+++ cypress/integration/WidgetsForEnrollmentPages/WidgetEnrollment/index.js
@@ -19,18 +19,18 @@ Then('the enrollment widget should be opened', () => {
 });
 
 Then('the user sees the enrollment date', () => {
-    cy.get('[data-test="widget-enrollment"]').within(() => {
+    cy.get('[data-test="widget-enrollment-enrollment-date"]').within(() => {
         cy.get('[data-test="widget-enrollment-icon-calendar"]').should('exist');
-        cy.get('[data-test="widget-enrollment-enrollment-date"]')
-            .contains(`Date of enrollment ${getCurrentYear()}-08-01`)
+        cy.get('[data-test="widget-enrollment-date"]')
+            .contains(`Date of enrollment: ${getCurrentYear()}-08-01`)
             .should('exist');
     });
 });
 
 Then('the user sees the incident date', () => {
-    cy.get('[data-test="widget-enrollment"]').within(() => {
-        cy.get('[data-test="widget-enrollment-incident-date"]')
-            .contains(`Date of birth ${getCurrentYear()}-08-01`)
+    cy.get('[data-test="widget-enrollment-incident-date"]').within(() => {
+        cy.get('[data-test="widget-enrollment-date"]')
+            .contains(`Date of birth: ${getCurrentYear()}-08-01`)
             .should('exist');
     });
 });