+++ src/core_modules/capture-core/components/Pages/EnrollmentEditEvent/EnrollmentEditEventPage.component.js
@@ -148,6 +148,7 @@ const EnrollmentEditEventPagePain = ({
                         teiId={teiId}
                         enrollmentId={enrollmentId}
                         programId={programId}
+                        readOnlyMode
                         onDelete={onDelete}
                         onAddNew={onAddNew}
                         onError={onEnrollmentError}