+++ src/core_modules/capture-core/components/WidgetEnrollmentEventNew/Validated/validated.types.js
@@ -25,5 +25,6 @@ export type Props = {|
     formRef: (formInstance: any) => void,
     dataEntryFieldRef: (instance: any, id: string) => void,
     rulesExecutionDependenciesClientFormatted: RulesExecutionDependenciesClientFormatted,
+    hiddenProgramStage: boolean,
     ...CssClasses,
 |};